requirements.txt
streamlit==1.48.1
pandas==2.3.2
plotly==6.3.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
numpy==2.3.2
matplotlib==3.10.5
seaborn==0.13.2
python-dotenv==1.1.1